+++ src/components/authentication/AccountDetails.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+
+import fire from '../../config/Fire';
+import AppDefaults from '../../AppDefaults';
+import * as UI from '../../controls/UI';
+// import './styles/SignUp.css';
+
+class AccountDetails extends Component {
+
+}
+
+export default AccountDetails;
