+++ src/components/authentication/SignUp.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 
-import fire from '../config/Fire';
-import AppDefaults from '../AppDefaults';
-import * as UI from '../controls/UI';
+import fire from '../../config/Fire';
+import AppDefaults from '../../AppDefaults';
+import * as UI from '../../controls/UI';
 import './styles/SignUp.css';
 
 const DialogType = Object.freeze({
