+++ src/components/main-interface/Home.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 
-import fire from '../config/Fire';
-import AppDefaults from '../AppDefaults';
-import * as UI from '../controls/UI';
+import fire from '../../config/Fire';
+import AppDefaults from '../../AppDefaults';
+import * as UI from '../../controls/UI';
 import './styles/Home.css';
 
 class Home extends Component {
