+++ src/components/main-interface/styles/Home.css
